Execute flight_tables_status_fix_mk1.sql
The original ELT job did not have the flight_status attribute in the airline_flights_scheduled table, and also filtered out cancelled and diverted flights.  This SQL process is lengthy, but it creates the airline_flights_scheduled table as "_new" with the flight_status attribute, and then populates the table, including the data from cancelled and diverted flight tables.  Then it renames the tables, recreates the PK, NK, FK objects with the right names, and recreates the view on top.